Getting rid of the driver class and re-arranging everything.
