+++ spring-cloud-gcp-samples/spring-cloud-gcp-data-spanner-repository-sample/README.adoc
@@ -1,0 +1,54 @@
+= Spring Cloud GCP Cloud Spanner Starter Example
+
+This code sample demonstrates how to read and write POJOs from Google Cloud Spanner using the Spring Data Cloud Spanner module link:../../spring-cloud-gcp-starters/spring-cloud-gcp-starter-data-spanner[Spring Cloud GCP Cloud Spanner Starter].
+
+== Running the example
+
+=== Running on GCP
+
+image:http://gstatic.com/cloudssh/images/open-btn.svg[link=https://ssh.cloud.google.com/cloudshell/editor?cloudshell_git_repo=https%3A%2F%2Fgithub.com%2FGoogleCloudPlatform%2Fspring-cloud-gcp&cloudshell_open_in_editor=spring-cloud-gcp-samples/spring-cloud-gcp-data-spanner-sample/README.adoc]
+
+Create a new Cloud Spanner instance named "spring-demo".
+The example application will create the database and tables if they do not already exist.
+
+https://cloud.google.com/sdk/gcloud/reference/auth/application-default/login[If you are authenticated in the Cloud SDK], your credentials will be automatically found by the Spring Boot Starter for Google Cloud Spanner.
+
+Alternatively, https://console.cloud.google.com/iam-admin/serviceaccounts[create a service account from the Google Cloud Console] and download its private key.
+Then, uncomment the `spring.cloud.gcp.spanner.credentials.location` property in the link:src/main/resources/application.properties[application.properties] file and fill its value with the path to your service account private key on your local file system, prepended with `file:`.
+
+=== Running  locally using emulator
+
+If you would like to run the sample locally using Cloud Spanner emulator, please follow https://cloud.google.com/spanner/docs/emulator[this steps] to set up and start the emulator.
+
+After that, you would need to create an instance using this command:
+
+----
+$ gcloud spanner instances create spring-demo --config=emulator-config --description="Test Instance" --nodes=1
+----
+
+Note that you need to set the `SPANNER_EMULATOR_HOST` environment variable to use the emulator.
+
+----
+$ export SPANNER_EMULATOR_HOST=localhost:9010
+----
+
+=== Running the application
+
+Run `$ mvn clean install` from the root directory of the project.
+
+Run the `SpannerRepositoryExample`  Spring Boot console apps.
+This causes the app to generate sample `Trade` objects filled with some demo data and persists them to the Cloud Spanner table created earlier.
+These POJOs are then retrieved and their contents are printed in the browser.
+
+.. To run the SpannerRepositoryExample:
+
+  `mvn spring-boot:run`
+
+Note that the sample application, still running, also provides REST endpoints.
+For example, you can retrieve all `Trade` objects in the repository by using
+`curl http://localhost:8080/trades`, or any specific trade via
+`curl http://localhost:8080/trades/<trader_id>,<trade_id>`.
+This is achieved using https://projects.spring.io/spring-data-rest/:[Spring Data REST] and the `@RepositoryRestResource` annotation.
+
+You can also write trades using
+`curl -XPOST -H"Content-Type: application/json" -d@test.json http://localhost:8080/trades/` where the file `test.json` holds the JSON representation of a `Trade` object.
