+++ spring-cloud-gcp-samples/spring-cloud-gcp-data-spanner-repository-sample/src/main/java/com/example/SpannerRepositoryExample.java
@@ -25,7 +25,10 @@ import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
+import org.springframework.boot.ApplicationRunner;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
 
 /**
  * Example repository usage.
@@ -34,7 +37,8 @@ import org.springframework.stereotype.Component;
  * @author Balint Pato
  * @author Mike Eltsufin
  */
-@Component
+
+@SpringBootApplication
 public class SpannerRepositoryExample {
 	private static final Log LOGGER = LogFactory.getLog(SpannerRepositoryExample.class);
 
@@ -122,5 +126,18 @@ public class SpannerRepositoryExample {
 			this.spannerDatabaseAdminTemplate.executeDdlStrings(Arrays.asList(
 					this.spannerSchemaUtils.getCreateTableDdlString(Trader.class)), true);
 		}
+	}
+
+	public static void main(String[] args) {
+		SpringApplication.run(SpannerRepositoryExample.class, args);
+	}
+
+	@Bean
+	ApplicationRunner applicationRunner() {
+		return args -> {
+			LOGGER.info("Running the Spanner Repository Example.");
+			runExample();
+
+		};
 	}
 }
